requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
uvloop>=0.21.0,<1.0.0
asyncpg>=0.30.0,<1.0.0
redis>=5.2.0,<6.0.0
discord.py>=2.4.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
cryptography>=44.0.0,<45.0.0
aiohttp>=3.11.0,<4.0.0
pydantic>=2.10.0,<3.0.0
httpx>=0.28.0,<1.0.0
python-multipart>=0.0.17
